viewmodel, backend

+++ immo_estimation/app/src/main/java/mastersid/richard/immo_estimation/data/EstimResult.kt
@@ -2,7 +2,7 @@ package mastersid.richard.immo_estimation.data
 
 sealed interface EstimResult {
     @JvmInline
-    value class Estimated(val v: Int ): EstimResult
+    value class Estimated(val v: Float ): EstimResult
 
     @JvmInline
     value class Failed(val text: String): EstimResult
